Move mode enum to src/types
