+++ src/types/scan.ts
@@ -1,0 +1,6 @@
+export enum Mode {
+  REGISTERING = "registering",
+  MANUAL = "manual",
+  MATCHING = "matching",
+  BLANK = "",
+}
